requirements_dev.txt
django-extensions ==0.6
django-debug-toolbar ==0.9.4